inference-worker/requirements.txt
redis
rq
langchain
langchain-openai
langchain-community
python-dotenv
openai
